pyproject.toml
[project]
name = "multi-research-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
]
